drop 3.6, update 3.7 to pypy7.3.6, add failing 3.8

+++ recipe/build.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -eo pipefail
+
 PYPY3_SRC_DIR=$SRC_DIR/pypy3
 
 if [[ "$target_platform" == "osx-64" ]]; then
@@ -59,7 +61,7 @@ mkdir -p $TARGET_DIR
 ${PYTHON} ./package.py --targetdir="$TARGET_DIR" --archive-name="$ARCHIVE_NAME"
 
 cd $TARGET_DIR
-tar -xvf $ARCHIVE_NAME.tar.bz2
+tar -xf $ARCHIVE_NAME.tar.bz2
 
 # Move all files from the package to conda's $PREFIX.
 cp -r $TARGET_DIR/$ARCHIVE_NAME/* $PREFIX
@@ -86,19 +88,35 @@ if [[ "$target_platform" == "linux"* ]]; then
     rm -f $PREFIX/bin/pypy3.debug
 fi
 
-# Move the generic file name to somewhere that's specific to pypy
-mv $PREFIX/README.rst $PREFIX/lib_pypy/
 # License is packaged separately
 rm $PREFIX/LICENSE
-
-# Make sure the site-packages dir match with cpython
 PY_VERSION=$(echo $PKG_NAME | cut -c 5-)
-mkdir -p $PREFIX/lib/python${PY_VERSION}/site-packages
-mv $PREFIX/site-packages/README $PREFIX/lib/python${PY_VERSION}/site-packages/
-rm -rf $PREFIX/site-packages
-ln -sf $PREFIX/lib/python${PY_VERSION}/site-packages $PREFIX/site-packages
 
-echo PWD is ${PWD}
+if [[ -d $PREFIX/lib_pypy ]]; then
+	# For pypy<3.8, the layout needs to be fixed up.
+
+	# Move the generic file name to somewhere that's specific to pypy
+	mv $PREFIX/README.rst $PREFIX/lib_pypy/
+	# Make sure the site-packages dir match with cpython
+	mkdir -p $PREFIX/lib/python${PY_VERSION}/site-packages
+	mv $PREFIX/site-packages/README $PREFIX/lib/python${PY_VERSION}/site-packages/
+	rm -rf $PREFIX/site-packages
+	ln -sf $PREFIX/lib/python${PY_VERSION}/site-packages $PREFIX/site-packages
+	pushd $PREFIX
+	pypy -m lib2to3.pgen2.driver lib-python/3/lib2to3/Grammar.txt
+	pypy -m lib2to3.pgen2.driver lib-python/3/lib2to3/PatternGrammar.txt
+	popd
+else
+	pushd $PREFIX
+	pypy -m lib2to3.pgen2.driver lib/pypy${PY_VERSION}/lib2to3/Grammar.txt
+	pypy -m lib2to3.pgen2.driver lib/pypy${PY_VERSION}/lib2to3/PatternGrammar.txt
+	popd
+fi
+
+echo sysconfig $(pypy -c "from distutils import sysconfig; print(sysconfig)")
+echo get_python_inc $(pypy -c "from distutils import sysconfig; print(sysconfig.get_python_inc())")
+echo INCLUDEPY $(pypy -c "from distutils import sysconfig; print(sysconfig.get_config_var('INCLUDEPY'))")
+ls $(pypy -c "from distutils import sysconfig; print(sysconfig.get_config_var('INCLUDEPY'))")
 # Build the c-extension modules for the standard library
 pypy -c "import _testcapi"
 pypy -c "import _ctypes_test"
@@ -106,12 +124,13 @@ pypy -c "import _testmultiphase"
 
 # Run the python stdlib tests
 timeout 60m pypy3 -m test --pgo -j${CPU_COUNT} || true;
-cd $PREFIX
-pypy -m lib2to3.pgen2.driver lib-python/3/lib2to3/Grammar.txt
-pypy -m lib2to3.pgen2.driver lib-python/3/lib2to3/PatternGrammar.txt
 
 
-cd $PREFIX/lib-python
-pypy3 -m compileall . || true;
-cd $PREFIX/lib_pypy
-pypy3 -m compileall . || true;
+if [[ -d $PREFIX/lib_pypy ]]; then
+	cd $PREFIX/lib-python
+	pypy3 -m compileall . || true;
+	cd $PREFIX/lib_pypy
+	pypy3 -m compileall . || true;
+else
+	pypy3 -m compileall $PREFIX/lib/pypy${PY_VERSION} || true;
+fi
